new file component refactor
