+++ libs/interpret-vision/src/lib/VisionExplanationDashboard/Controls/FlyoutObjectDetection.tsx
@@ -20,6 +20,7 @@ import { FluentUIStyles } from "@responsible-ai/core-ui";
 import { localization } from "@responsible-ai/localization";
 import * as React from "react";
 import { CanvasTools } from "vott-ct";
+import { DetectionDetails } from "./DetectionDetails";
 
 import * as FlyoutStyles from "../utils/FlyoutUtils";
 import { getObjectDetectionImageAltText } from "../utils/getAltTextUtils";
@@ -123,7 +124,7 @@ export class FlyoutObjectDetection extends React.Component<
                   verticalAlign="center"
                 >
                   <Stack.Item>
-                    <FlyoutODUtils.DetectionDetails
+                    <DetectionDetails
                       item={item}
                       correctDetections={correctDetections}
                       incorrectDetections={incorrectDetections}
