+++ libs/interpret-vision/src/lib/VisionExplanationDashboard/Controls/FlyoutObjectDetectionUtils.tsx
@@ -28,55 +28,12 @@ export interface IFlyoutState {
   editorCallback?: HTMLDivElement;
 }
 
-interface IDetectionDetailsProps {
-  item: IVisionListItem; // replace with actual type
-  correctDetections: string;
-  incorrectDetections: string;
-}
-
 export const stackTokens = {
   large: { childrenGap: "l2" },
   medium: { childrenGap: "l1" }
 };
 
 const theme = getTheme();
-
-export class DetectionDetails extends React.Component<IDetectionDetailsProps> {
-  public render(): React.ReactNode {
-    return (
-      <Stack
-        tokens={stackTokens.large}
-        horizontalAlign="start"
-        verticalAlign="start"
-      >
-        <Stack
-          horizontal
-          tokens={stackTokens.medium}
-          horizontalAlign="center"
-          verticalAlign="center"
-        />
-        <Stack.Item>
-          <Text variant="large">
-            {localization.InterpretVision.Dashboard.indexLabel}
-            {this.props.item?.index}
-          </Text>
-        </Stack.Item>
-        <Stack.Item>
-          <Text variant="large">
-            {localization.InterpretVision.Dashboard.correctDetections}
-            {this.props.correctDetections}
-          </Text>
-        </Stack.Item>
-        <Stack.Item>
-          <Text variant="large">
-            {localization.InterpretVision.Dashboard.incorrectDetections}
-            {this.props.incorrectDetections}
-          </Text>
-        </Stack.Item>
-      </Stack>
-    );
-  }
-}
 
 export class ColorLegend extends React.Component {
   public render(): React.ReactNode {
